requirements.txt
flask==2.0.3
werkzeug==2.0.3
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.46
flask-migrate==3.1.0
flask-jwt-extended==4.4.4
psycopg2-binary==2.9.6
numpy==1.22.4
pandas==1.4.3
xlsxwriter==3.0.9
python-docx==0.8.11
gunicorn==20.1.0